requirements.txt
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
PyQt5>=5.15.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
numpy>=1.19.0
dlib==19.24.2
face-recognition==1.3.0
Pillow==9.0.1
PyYAML>=5.4.0
sqlite3
pyserial>=3.5
